Use struct to represent final state of CD player
Previous isValidCDPlayerString() kept track of the CD player's state,
such as whether it's open. Although necessary to validate
cd_player_actions, that state info is lost since only a boolean is
returned.

Since state info is useful for the other functions (required for this
assignment), isValidCDPlayerString() is renamed to getCdPlayerState()
and rewritten to return a struct containing the state info.
isValidCDPlayerString() is refactored to use the struct data,
specifically the boolean is_valid.